Current state of a CMake build system
The CMakeLists.txt file doesn't do final linking and all yet
because I couldn't get the MSVC compiler to compile the boost
python code without ICE'ing, which according to google is a very
well known bug in the version of MSVC's compiler that is used for
this project, so if that can ever be upgraded then it should fix
it.

Theoretically the CMakeLists.txt file should work on windows as
well, the linux stuff is encompassed into a dedicated toolchain
that should be used there instead.

Other than linking not being done yet for the project, the
CMakeLists.txt file also doesn't have Unity builds enabled on it
yet either as I wanted to get a full normal compile to succeed
first.

The C2C source has a *lot* of missing headers, that only
incidentally worked before because the fastbuild unity build was
combining files in such a way that it worked, but the issue then
is that if the Unity build sorting changes, such as by adding or
removing a source file, then it can break.  Normal builds should
always work properly and correctly before ever enabling Unity
builds.  In addition, unity builds mean that incremental
compilation, such as for debug builds, becomes incredibly more
costly, so Unity builds are usually restricted to Release builds.